Revert "Core/DBLayer: Correct few more wrong read types. Step2"
This reverts commit 695cd1e693525e9ba4e33f006cce9794022bba1c.
DO NOT COMMIT UNFINISHED STUFF.

+++ src/server/game/Achievements/AchievementMgr.cpp
@@ -646,7 +646,7 @@ void AchievementGlobalMgr::LoadAchievementCriteriaData()
             continue;
         }
 
-        uint32 dataType = fields[1].GetUInt8();
+        uint32 dataType = fields[1].GetUInt32();
         const char* scriptName = fields[4].GetCString();
         uint32 scriptId = 0;
         if (strcmp(scriptName, "")) // not empty
